requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
html5lib==1.0.1
idna==2.9
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
pdoc3==0.8.1
plac==1.1.3
praw==6.5.1
prawcore==1.0.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
update-checker==0.16
urllib3==1.25.9
webencodings==0.5.1
websocket-client==0.57.0
